requirements.txt
# ── Core data ──────────────────────────────────────────────
pandas==2.2.3
pyarrow==16.1.0
numpy==1.26.4

# ── Database ───────────────────────────────────────────────
mysql-connector-python==9.0.0
SQLAlchemy==2.0.35

# ── API ────────────────────────────────────────────────────
fastapi==0.115.5
uvicorn==0.30.6
pydantic==2.9.2
httpx==0.27.2

# ── Visualization ──────────────────────────────────────────
streamlit==1.39.0

# ── Orchestration ──────────────────────────────────────────
apache-airflow==2.10.3

# ── DevOps / Containers ────────────────────────────────────
# (Docker se instala en el devcontainer, no aquí)

# ── Notebooks ──────────────────────────────────────────────
jupytext==1.16.4
jupyter==1.1.1
ipykernel==6.29.5

# ── Utilities ──────────────────────────────────────────────
python-dotenv==1.0.1
requests==2.32.3
